move `ImmutableRecord` to its own file
